% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{BindMatrices}
\alias{BindMatrices}
\title{Function to bind matrices and return the necessary outputs}
\usage{
BindMatrices(cluster_list, correlation_matrix)
}
\arguments{
\item{correlation_matrix}{}
}
\description{
Function to bind matrices and return the necessary outputs
}
