% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{CombinedPar}
\alias{CombinedPar}
\title{Finds the difference between Euclidean Distance and Spearman Dissimilarity}
\usage{
CombinedPar(allptmtable.df, allptmtable)
}
\arguments{
\item{PTM}{data set & data frame}
}
\value{
TSNE coordinates
}
\description{
Finds the difference between Euclidean Distance and Spearman Dissimilarity
}
\examples{
CombinedPar(allptmtable.df, allptmtable)
}
