% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{CorrelationNetwork}
\alias{CorrelationNetwork}
\title{Function to create a correlation network}
\usage{
CorrelationNetwork(bind_result)
}
\arguments{
\item{bind_result}{}
}
\description{
Function to create a correlation network
}
