% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{EuclideanDistance}
\alias{EuclideanDistance}
\title{Calculates Euclidean Distance between}
\usage{
EuclideanDistance(allptmtable.df)
}
\arguments{
\item{Post}{translation modification data frame}
}
\value{
TSNE Euclidean Distance
}
\description{
Calculates Euclidean Distance between
}
\examples{
(EuclideanDistance(allptmtable.df)
}
