% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{FindCommonCluster}
\alias{FindCommonCluster}
\title{Finds correlations between clusters}
\usage{
FindCommonCluster(
  eu_allptms_tsne,
  sp_allptms_tsne,
  sed_allptms_tsne,
  allptmtable_df,
  output_dir = "plots"
)
}
\arguments{
\item{output_dir}{}
}
\value{
list(eu_allptms_list = eu_allptms_list, sp_allptms_list = sp_allptms_list, sed_allptms_list = sed_allptms_list
}
\description{
Finds correlations between clusters
}
