% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{MakeAdjMatrix}
\alias{MakeAdjMatrix}
\title{Creates an adjacency matrix from the given data set}
\usage{
MakeAdjMatrix(list.element)
}
\arguments{
\item{list.element}{}
}
\value{
list.el.mat
}
\description{
Creates an adjacency matrix from the given data set
}
