% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{MakeClusterList}
\alias{MakeClusterList}
\title{Makes a list of cluster groupings}
\usage{
MakeClusterList(tsnedata, toolong, tbl.sc)
}
\arguments{
\item{TSNE}{Data, Cluster Distance Number, TBD}
}
\value{
TSNE list (filtered)
}
\description{
Makes a list of cluster groupings
}
