% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{process_ptms_data}
\alias{process_ptms_data}
\title{Title}
\usage{
process_ptms_data(eu.sp.sed.allptms, sed.allptms.peps, AlldataPTMs_cor)
}
\arguments{
\item{AlldataPTMs_cor}{}
}
\description{
Title
}
