% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{SpearmanDissimilarity}
\alias{SpearmanDissimilarity}
\title{Calculates Spearman dissimilarity and t-SNE from a given data set}
\usage{
SpearmanDissimilarity(allptmtable)
}
\arguments{
\item{Post}{translation modifications table}
}
\value{
Spearman Dissimilarity
}
\description{
Calculates Spearman dissimilarity and t-SNE from a given data set
}
\examples{
(SpearmanDissimilarity(allptmtable)
}
