% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{CombinedPar}
\alias{CombinedPar}
\title{Combines Spearman dissimilarity and Euclidean distance in parallel}
\usage{
CombinedPar(allptmtable.df, allptmtable)
}
\arguments{
\item{allptmtable.df}{A data frame containing numeric data.}

\item{allptmtable}{A dataset for post-translational modifications.}
}
\value{
A matrix containing t-SNE coordinates (3D).
}
\description{
This function uses parallel computing to calculate both Spearman dissimilarity and
Euclidean distance, combines them, and performs t-SNE.
}
\examples{
CombinedPar(allptmtable.df, allptmtable)
}
